chore: commit all current workspace changes

+++ internal/delivery/http/dto/cart/cart_dto.go
@@ -12,17 +12,17 @@ type UpdateShoppingCartItemRequest struct {
 }
 
 type ShoppingCartItemResponse struct {
-	ID        string  `json:"id"`
-	ProductID string  `json:"product_id"`
-	Quantity  int     `json:"quantity"`
-	PriceBTC  float64 `json:"price_btc"`
-	Subtotal  float64 `json:"subtotal"`
+	ID        string `json:"id"`
+	ProductID string `json:"product_id"`
+	Quantity  int    `json:"quantity"`
+	PriceBTC  int64  `json:"price_btc"`
+	Subtotal  int64  `json:"subtotal"`
 }
 
 type ShoppingCartResponse struct {
 	ID        string                     `json:"id"`
 	UserID    string                     `json:"user_id"`
-	TotalBTC  float64                    `json:"total_btc"`
+	TotalBTC  int64                      `json:"total_btc"`
 	Items     []ShoppingCartItemResponse `json:"items"`
 	CreatedAt string                     `json:"created_at,omitempty"`
 	UpdatedAt string                     `json:"updated_at,omitempty"`
