+++ internal/delivery/http/dto/order/order_dto.go
@@ -7,7 +7,7 @@ type OrderResponse struct {
 	UserID    string                         `json:"user_id"`
 	AddressID string                         `json:"address_id"`
 	Status    string                         `json:"status"`
-	TotalBTC  float64                        `json:"total_btc"`
+	TotalBTC  int64                          `json:"total_btc"`
 	Items     []dorderitem.OrderItemResponse `json:"items"`
 	CreatedAt string                         `json:"created_at,omitempty"`
 	UpdatedAt string                         `json:"updated_at,omitempty"`
